api: make active views a core concept
The active view should be set/unset according to the global keyboard
focus, not according to what the active view on the output was.

In addition, output->focus_view() will no longer raise the view to the
top by default.

There is also a new workspace_manager function, restack_above()