input: do not send internal events for artificial touch motion
When a surface is mapped/unmapped, we need to update the touch focus
surface. We do this by generating an aritifial motion event. However,
this shouldn't be propagated to plugins, otherwise we can get nasty
races where a plugin maps/unmaps a surface, this triggers motion, which
triggers map/unmap in the same plugin and so on.